refactor: fix image path to asserts/images
